refactor(reader): extract column buffer skip logic
